Reduce load on PCA9685 handling
- filter PWM updates based on value that would be communicated to device.
- should reduce updates by a factor of 10 at least

+++ outputs/platforms/torci2cpca9685.h
@@ -21,7 +21,7 @@ class TorcI2CPCA9685 : public TorcI2CDevice
     ~TorcI2CPCA9685();
 
   protected:
-    bool                   SetPWM (int Channel, double Value);
+    bool                   SetPWM (int Channel, int Value);
 
   private:
     TorcI2CPCA9685Channel *m_outputs[16];
